chore: Remove lib.rs and move behavior tests from integration to unit tests
