+++ .github/workflows/test_behavior.yml
@@ -31,7 +31,7 @@ jobs:
           echo "MinIO not ready" >&2; exit 1
 
       - name: Run Behavior Tests
-        run: cargo test --test behavior --quiet
+        run: cargo test tests::behavior -- --quiet
         env:
           RUST_LOG: info
