+++ Cargo.toml
@@ -23,7 +23,7 @@ clap = { version = "4.5.47", features = ["derive", "env"] }
 dialoguer = "0.12.0"
 futures = "0.3.30"
 log = "0.4.28"
-opendal = { version = "0.54.0" }
+opendal = { version = "0.54.0", default-features = false }
 snafu = "0.8.9"
 tokio = { version = "1.47.1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -50,8 +50,3 @@ predicates = "3.1.3"
 rand = "0.9.2"
 tracing-subscriber = { version = "0.3.20", features = ["env-filter", "tracing-log"] }
 tempfile = "3.14.0"
-
-[[test]]
-harness = false
-name = "behavior"
-path = "tests/behavior/main.rs"
