+++ src/cli/mod.rs
@@ -4,6 +4,4 @@ pub mod entry;
 pub mod prompts;
 pub mod storage;
 
-pub use context::CliContext;
-pub use entry::{Args, Command, ConfigCommand, GlobalOptions, run, run_with_prompt};
-pub use prompts::Prompt;
+pub use entry::{Args, run};
