--- src/lib.rs
@@ -1,5 +1,0 @@
-pub mod cli;
-pub mod config;
-pub mod error;
-pub mod storage;
-pub mod utils;
