--- tests/data/small.txt
@@ -1,1 +1,0 @@
-This is a small test file for storify.
