--- tests/data/special_dir !@#$%^&()_+-=;'/file_in_special_dir.txt
@@ -1,1 +1,0 @@
-content from a file in a special directory
